machine: add property param group
Instead of filtering the static properties in the UI do that in the machine.
This is a preparation for actually saving the values in a song. It also makes
that part testable removes non ui code from the ui. In addition we can avoid
showing the dialog if there are no settings.