cmd-pattern-control-source: add a new control-source to handle cmd-patterns
This new cs is controlling the machine::state property based on the cmd-
patterns. To achieve that it is attached to the mute property of the gain
elements inside the machine. Since this will trigger notify signals from the
render threads we need to dispatch them to the main_context.