signals: use the wonderful g_signal_connect_object
When we connect to a long living object from a short living object use this api,
as it will remove the signal handler when we finalize the short living object.
This simplifies our destruction code.

Also add a bt_g_signal_connect_object() variant so that we can apply this in a
consistent way.