audiosynth: add proper discont handling
Finish what we started with commit d4fc4c04. Add a new 'reset' vmthod.
When generating a discont buffer (e.g. when seeking), this is called.
Elements can reset the sound generation (cut running notes).

The implementation won't generate a discont on the first seek to not
reset properties set on creation. This fixes running commands via
gst-launch, but is just another issue wrt not being able to configure
a playback segment.