registry: explicity activate the discovery modules.
We instantiated and started them from the registry constructor. This causes them
to also start when e.g. building gobject instrospection and gtk-docs. Also some
code might want to use the registry, but not the discovery modules (e.g. a pure
player).

This commit add new api to explicity start the discoverers and internal api,
that discoverers can use to check if there is an active registry. The later
allows to instantiate the discoverers in a way that the won't do anything.

Fixes #69