core: add a local install helper
When building from sources and installing to eg $HOME or /usr/local one needs
to export a bunch of env-vars. This is a PITA for new users.

First, update the docs and tell what vars one needs to setup for what reason.

Next, use the new helper in buzztrax-{cmd,edit}. This checks the $PREFIX and
modifies the envvars as needed.

See #78