Fixed sequence patterns being lost when setting song end
When the "song end" was set, patterns were lost and even undo didn't retrieve them. If you accidentally set the end of
the song at tick 0, then the whole song was lost.

I also changed the behaviour here; the sequence is no longer truncated when setting the end point earlier in the song.
The sequence is only truncated up to the point where only empty patterns are in the sequence. This behaviour matches
Buzz and I think it's kinder to users, too.

C99 mode enabled on Stefan's suggestion.